+++ server/tests/onlyUsers.test.js
@@ -1,4 +1,4 @@
-const { getOnlyUsersIDs, getUserPostInfo, getUserCommentsInfo } = require('../src/utils/onlyUsers');
+const { getOnlyUsersIDs, getUserPostInfo, getUserCommentsInfo } = require('../utils/onlyUsers');
 
 const mockComments = [{
   _id: '2',
